version bump -> 0.2.3

+++ app/download_recommendations.py
@@ -10,7 +10,8 @@ from utils import retrieve_release_from_github
 settings = Settings()
 
 # first, clean the recommendation directory if it exists
-shutil.rmtree(settings.recommendation_path)
+if settings.recommendation_path.exists():
+    shutil.rmtree(settings.recommendation_path)
 
 retrieve_release_from_github(
     repository_url=settings.gh_repository, target_path=settings.recommendation_path
